// shared/src/types/LevelDefinition.d.ts
export interface SlopeggleLevelOverrides {
    /** Number of peg rows to generate */
    pegRows?: number;
    /** Total number of orange (target) pegs */
    orangePegCount?: number;
    /** Include the free-ball bucket hazard */
    hasBucket?: boolean;
    /** Include teleport portals */
    hasPortals?: boolean;
    /** World width in meters */
    worldWidth?: number;
    /** World height in meters */
    worldHeight?: number;
    /** Ball launch force multiplier */
    launchForceMultiplier?: number;
    /** Enable slow-motion on final peg */
    dramaticFinalPeg?: boolean;
}